Fixes the bug that framebuffer is not reset properly.
This commit fixes the `Widget::RenderDefaultFramebuffer()` method that the `default_framebuffer_` is not reset properly, and that can cause a problem when the size of the framebuffer changed. This commit also replaces `int` type with `float` type for variables keeping widget's dimension.

+++ moui/widgets/button.cc
@@ -83,7 +83,6 @@ Button::Button()
   title_label_->set_text_horizontal_alignment(Label::Alignment::kCenter);
   title_label_->set_text_vertical_alignment(Label::Alignment::kMiddle);
   AddChild(title_label_);
-
 }
 
 Button::~Button() {
@@ -189,8 +188,8 @@ void Button::Render(NVGcontext* context) {
   if (*framebuffer == nullptr)
     return;
 
-  const int kWidth = GetWidth();
-  const int kHeight = GetHeight();
+  const float kWidth = GetWidth();
+  const float kHeight = GetHeight();
   nvgBeginPath(context);
   nvgRect(context, 0, 0, kWidth, kHeight);
   NVGpaint paint = nvgImagePattern(context, 0, kHeight, kWidth, kHeight, 0,
